Include asserts into the base class to prevent not handling it
